fixes and refactoring

+++ domain/user_repository.go
@@ -4,6 +4,6 @@ type UserRepository interface {
 	CreateUser(user *User) error
 	GetUserByUsername(username string) (*User, error)
 	GetUserByEmail(email string) (*User, error)
-	GetUserByID(id int64) (*User, error)
+	GetUserByID(id int) (*User, error)
 	UpdateUser(user *User) error
 }
